add world object loading and more data driven design
- changed some u32 to usize because why are they only u32...
- better data reading can now read in world objects
- databases will have get_by_x where x can be id or name
- cleaning and abstracting some code

+++ src/mining.rs
@@ -83,7 +83,7 @@ impl<'a> System<'a> for DamageSystem {
 
             // Addition is used because damage dealt can be positive or negative
             let new_hp = stats.hp as i32 + damage_dealt;
-            stats.hp = if new_hp >= 0 { new_hp as u32 } else { 0 };
+            stats.hp = if new_hp >= 0 { new_hp as usize } else { 0 };
 
             println!(
                 "Old HP: {} | Damage Dealt: {} | New HP: {}",
